fixed a bug where we couldn't deserialize 64 bits max value

+++ include/ManiZ/Json.h
@@ -213,7 +213,7 @@ namespace ManiZ
 
 	namespace _impl
 	{
-		using json_variant = std::variant<long, unsigned long, double, std::string, bool>;
+		using json_variant = std::variant<long long, unsigned long long, double, std::string, bool>;
 	}
 
 	class JsonObject
@@ -240,7 +240,7 @@ namespace ManiZ
 		requires std::is_integral_v<T> && std::is_unsigned_v<T>
 		T get() const
 		{
-			return static_cast<T>(std::get<unsigned long>(m_value));
+			return static_cast<T>(std::get<unsigned long long>(m_value));
 		}
 
 		template<typename T>
@@ -248,11 +248,11 @@ namespace ManiZ
 		T get() const
 		{
 			// we deserialize to long only if the value is negative, otherwise it is set to the unsigned long slot of the variant.
-			if (const long* value = std::get_if<long>(&m_value))
+			if (const long long* value = std::get_if<long long>(&m_value))
 			{
 				return static_cast<T>(*value);
 			}
-			return static_cast<T>(std::get<unsigned long>(m_value));
+			return static_cast<T>(std::get<unsigned long long>(m_value));
 		}
 
 		template<typename T>
@@ -468,7 +468,7 @@ namespace ManiZ
 					}
 					else
 					{
-						return JsonObject(std::stoul(value));
+						return JsonObject(std::stoull(value));
 					}
 				}
 			}
